+++ packages/typoas-runtime/src/auth/index.ts
@@ -3,3 +3,4 @@ export * from './api-key-auth';
 export * from './http-auth-basic';
 export * from './http-auth-bearer';
 export * from './oauth2-auth';
+export * from './open-id-connect-auth';